Change .ready state to match that of KotlinAudio and RNTP.
Before the .ready state was entered into after the player state changed to AVPlayer.Status.readyToPlay which indicates the player is ready to media. See https://developer.apple.com/documentation/avfoundation/avplayer/status/readytoplay.

In react-native-track-player we use this for `State.Ready`: “State indicating that the player is ready to start playing”

Because these functionalities do not match up and because we want to match the behavior of KotlinAudio, the `.ready` state is now entered into when the `AVPlayerItem.isPlaybackLikelyToKeepUp` property changes to true.