fix: #42 - update current item correctly upon removal
The previous code used to not update the current item correctly
when the current item was removed. This led to cases where you would
remove an item, then act on what you thought was the new item (e.g.
by trying to get its duration, by playing it, etc.) when AVPlayer
was still operating on the old item.

This fixes that by sending the right delegate notification when the
current item is the one being removed.