Ensure order of emitted events
Global dispatch queues in iOS are concurrent queues, which means tasks are started in the order they were added, but they may not finish in that order. This is because concurrent queues allow multiple tasks to run at the same time.

Serial queues run tasks one at a time in the order that they were added. This means that if you dispatch tasks to this queue, they will be executed in the order they were added (or dispatched) to the queue.